Fix implement Bloc

+++ lib/core/features/matriculacion_vehicular/domain/entities/info_vehiculo_entity.dart
@@ -37,7 +37,7 @@ class InfoVehiculoEntity extends Equatable {
   final String mensajeMotivoAuto;
   final String placa;
   final String camvCpn;
-  final int cilindraje;
+  final double cilindraje;
   final int fechaCompra;
   final int anioUltimoPago;
   final String marca;
